fix(db): use BIGINT for Long JPA ID columns AB#1929 AB#1930

+++ backend/src/main/resources/db/migration/V003__create_notification.sql
@@ -7,7 +7,7 @@
 -- Notification transactional table for email records
 CREATE TABLE notification (
     id BIGINT IDENTITY(1,1) PRIMARY KEY,
-    program_id INT NOT NULL,
+    program_id BIGINT NOT NULL,
     notification_type NVARCHAR(50) NOT NULL,
     recipient_email NVARCHAR(200) NOT NULL,
     subject NVARCHAR(500) NOT NULL,
